\name{GxTheta}
\alias{GxTheta}
\alias{GxTheta}
\docType{package}
\title{
This software implements a statistical approach to using ancestry in gene-by-gene (GxG) and gene-by-environement (GxE) interaction analyses.
}
\description{
Provides a simulation framework for power analysis. 
}
\details{
\tabular{ll}{
Package: \tab GxTheta\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-12-10\cr
License: \tab GPL-2 \cr
}
}
\author{
Danny S. Park, Eric R. Gamazon, Noah Zaitlen
Maintainer: Danny S. Park <sihyung@gmail.com>
}
\references{
An ancestry based approach for detecting interactions by Park DS et al. Submitted 
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ CalculatePowerGxE(save.path, pval.cutoff, E.effect.size, genetic.effect.size, global.anc.effect, interaction.effect.size, n.idvs, sigma.max, sigma.step, n.iter) ~~
~~ SimulateDifferentialTagging(save.path, n.iter, genetic.effect.size, n.idvs) ~~
}
